+++ ssas/logger.go
@@ -1,10 +1,14 @@
 package ssas
 
 import (
+	"context"
+	"fmt"
+	"net/http"
 	"os"
 	"time"
 
 	"github.com/CMSgov/bcda-ssas-app/ssas/constants"
+	"github.com/go-chi/chi/v5/middleware"
 	"github.com/sirupsen/logrus"
 )
 
@@ -50,6 +54,48 @@ func init() {
 
 type APILoggerEntry struct {
 	Logger logrus.FieldLogger
+}
+
+func (l *APILoggerEntry) Write(status int, bytes int, header http.Header, elapsed time.Duration, extra interface{}) {
+	l.Logger = l.Logger.WithFields(logrus.Fields{
+		"resp_status": status, "resp_bytes_length": bytes,
+		"resp_elapsed_ms": float64(elapsed.Nanoseconds()) / 1000000.0,
+	})
+
+	l.Logger.Infoln("request complete")
+}
+
+func (l *APILoggerEntry) Panic(v interface{}, stack []byte) {
+	l.Logger = l.Logger.WithFields(logrus.Fields{
+		"stack": string(stack),
+		"panic": fmt.Sprintf("%+v", v),
+	})
+}
+
+// type to create context.Context key
+type CtxLoggerKeyType string
+
+const CtxLoggerKey CtxLoggerKeyType = "ctxLogger"
+
+// context.Context key to set/get logrus.FieldLogger value within request context
+
+// Gets the logrus.FieldLogger from a context
+func GetCtxLogger(ctx context.Context) logrus.FieldLogger {
+	entry := ctx.Value(CtxLoggerKey).(*APILoggerEntry)
+	return entry.Logger
+}
+
+// Gets the logrus.APILoggerEntry from a context
+func GetCtxEntry(ctx context.Context) *APILoggerEntry {
+	entry := ctx.Value(CtxLoggerKey).(*APILoggerEntry)
+	return entry
+}
+
+// Appends additional logrus.Field to a logrus.FieldLogger within a context
+func SetCtxEntry(r *http.Request, key string, value interface{}) {
+	if entry, ok := r.Context().Value(middleware.LogEntryCtxKey).(*APILoggerEntry); ok {
+		entry.Logger = entry.Logger.WithField(key, value)
+	}
 }
 
 func mergeNonEmpty(data Event) logrus.FieldLogger {
