+++ ssas/service/public/api.go
@@ -14,7 +14,6 @@ import (
 	"net/http"
 	"strconv"
 
-	"github.com/CMSgov/bcda-ssas-app/log"
 	"github.com/CMSgov/bcda-ssas-app/ssas"
 	"github.com/CMSgov/bcda-ssas-app/ssas/constants"
 	"github.com/CMSgov/bcda-ssas-app/ssas/service"
@@ -70,7 +69,7 @@ func ResetSecret(w http.ResponseWriter, r *http.Request) {
 	setHeaders(w)
 
 	if rd, err = readRegData(r); err != nil || rd.GroupID == "" {
-		log.Logger.Println("missing or invalid GroupID")
+		ssas.Logger.Println("missing or invalid GroupID")
 		service.JSONError(w, http.StatusUnauthorized, http.StatusText(http.StatusUnauthorized), "")
 		return
 	}
@@ -81,7 +80,7 @@ func ResetSecret(w http.ResponseWriter, r *http.Request) {
 	}
 
 	if err = json.Unmarshal(bodyStr, &req); err != nil {
-		log.SetCtxEntry(r, "bodyStr", bodyStr)
+		ssas.SetCtxEntry(r, "bodyStr", bodyStr)
 		service.JSONError(w, http.StatusBadRequest, "invalid_client_metadata", "Request body cannot be parsed")
 		return
 	}
@@ -142,7 +141,7 @@ func RegisterSystem(w http.ResponseWriter, r *http.Request) {
 	setHeaders(w)
 
 	if rd, err = readRegData(r); err != nil || rd.GroupID == "" {
-		log.Logger.Println("missing or invalid GroupID")
+		ssas.Logger.Println("missing or invalid GroupID")
 		// Specified in RFC 7592 https://tools.ietf.org/html/rfc7592#page-6
 		service.JSONError(w, http.StatusUnauthorized, http.StatusText(http.StatusUnauthorized), "")
 		return
@@ -157,7 +156,7 @@ func RegisterSystem(w http.ResponseWriter, r *http.Request) {
 
 	err = json.Unmarshal(bodyStr, &reg)
 	if err != nil {
-		log.SetCtxEntry(r, "bodyStr", bodyStr)
+		ssas.SetCtxEntry(r, "bodyStr", bodyStr)
 		service.JSONError(w, http.StatusBadRequest, "invalid_client_metadata", "Request body cannot be parsed")
 		return
 	}
@@ -246,13 +245,13 @@ func token(w http.ResponseWriter, r *http.Request) {
 
 	system, err := ssas.GetSystemByClientID(r.Context(), clientID)
 	if err != nil {
-		log.Logger.Errorf("The client id %s is invalid", err.Error())
+		ssas.Logger.Errorf("The client id %s is invalid", err.Error())
 		service.JSONError(w, http.StatusUnauthorized, http.StatusText(http.StatusUnauthorized), "invalid client id")
 		return
 	}
 	err = ValidateSecret(system, secret, r)
 	if err != nil {
-		log.Logger.Error("The client id and secret cannot be validated: ", err.Error())
+		ssas.Logger.Error("The client id and secret cannot be validated: ", err.Error())
 		service.JSONError(w, http.StatusUnauthorized, http.StatusText(http.StatusUnauthorized), err.Error())
 		return
 	}
@@ -260,7 +259,7 @@ func token(w http.ResponseWriter, r *http.Request) {
 	trackingID := uuid.NewRandom().String()
 
 	data, err := ssas.XDataFor(r.Context(), system)
-	log.Logger.Infof("public.api.token: XDataFor(%d) returned '%s'", system.ID, data)
+	ssas.Logger.Infof("public.api.token: XDataFor(%d) returned '%s'", system.ID, data)
 	if err != nil {
 		service.JSONError(w, http.StatusUnauthorized, http.StatusText(http.StatusUnauthorized), "no group for system")
 		return
@@ -372,7 +371,7 @@ func tokenV2(w http.ResponseWriter, r *http.Request) {
 	}
 
 	data, err := ssas.XDataFor(r.Context(), system)
-	log.Logger.Infof("public.api.token: XDataFor(%d) returned '%s'", system.ID, data)
+	ssas.Logger.Infof("public.api.token: XDataFor(%d) returned '%s'", system.ID, data)
 	if err != nil {
 		service.JSONError(w, http.StatusUnauthorized, http.StatusText(http.StatusUnauthorized), "no group for system")
 		return
@@ -483,7 +482,7 @@ func introspect(w http.ResponseWriter, r *http.Request) {
 	var answer = make(map[string]bool)
 	answer["active"] = true
 	if err = tokenValidity(reqV["token"], "AccessToken"); err != nil {
-		log.Logger.Infof("token failed tokenValidity")
+		ssas.Logger.Infof("token failed tokenValidity")
 		answer["active"] = false
 	}
 
@@ -514,12 +513,12 @@ func validateAndParseToken(w http.ResponseWriter, r *http.Request) {
 	var response = make(map[string]interface{})
 
 	if err := tokenValidity(tokenS, "AccessToken"); err != nil {
-		log.Logger.Infof("token failed tokenValidity")
+		ssas.Logger.Infof("token failed tokenValidity")
 		response["valid"] = false
 	} else {
 		claims := jwt.MapClaims{}
 		if _, _, err := new(jwt.Parser).ParseUnverified(tokenS, claims); err != nil {
-			log.Logger.Infof("could not unmarshal access token")
+			ssas.Logger.Infof("could not unmarshal access token")
 			service.JSONError(w, http.StatusInternalServerError, http.StatusText(http.StatusInternalServerError), "internal server error")
 			return
 		}
@@ -528,7 +527,7 @@ func validateAndParseToken(w http.ResponseWriter, r *http.Request) {
 		response["system_data"] = claims["system_data"]
 		sys, err := ssas.GetSystemByID(r.Context(), claims["sys"].(string))
 		if err != nil {
-			log.Logger.Infof("could not get system id")
+			ssas.Logger.Infof("could not get system id")
 			service.JSONError(w, http.StatusInternalServerError, http.StatusText(http.StatusInternalServerError), "internal server error")
 			return
 		}
